requirements.txt
build==0.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
docopt==0.6.2
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
packaging==22.0
pep517==0.13.0
pip-tools==6.12.1
pipreqs==0.4.11
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytils==0.4.1
requests==2.28.1
six==1.16.0
tomli==2.0.1
tools==0.1.9
urllib3==1.26.13
Werkzeug==2.2.2
yarg==0.1.9